requirements.txt
discord-py-slash-command ~= 2.2.0
discord.py ~= 1.7.3
requests ~= 2.25.1
